+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Next
 
+## 2.0.1
+
+- New `platform` name in 2.x.x - Refer to README.md
+- Improved warning on events for unknown group addresses
+- Fixed error in platform name in config schema
+
 ## 2.0.0
 
 - Improved error handling